requirements.txt
# Core dependencies
python-dotenv>=0.19.0
streamlit>=1.22.0
web3>=6.0.0
pillow>=9.0.0
pydantic>=1.10.8
asyncio>=3.4.3
requests>=2.28.0

# IPFS and blockchain
ipfshttpclient>=0.8.0
eth-account>=0.5.9
py-solc-x>=1.1.1

# AI and ML dependencies
torch>=2.0.0
torchvision>=0.15.0
ftfy>=6.1.1
regex>=2022.3.15
tqdm>=4.65.0
clip @ git+https://github.com/openai/CLIP.git
pytesseract>=0.3.10
opencv-python>=4.7.0
scipy>=1.10.1

# Web and API
fastapi>=0.95.1
uvicorn>=0.22.0
aiohttp>=3.8.4

# Storage
pyarrow>=12.0.0